fix(web): stabilize business panel resize gestures
